Funcionando. Só falta descobrir como...

+++ projetosopencl/multiplicacaodematriz/part1.cl
@@ -1,16 +1,15 @@
 #define STRINGIFY(A) #A
 std::string kernel_source = STRINGIFY(
 
-float function_example(float a, float b)
+int function_example(int a, int b)
 {
     return a + b;
 }
 
-__kernel void part1(__global float* a, __global float* b, __global float* c)
+__kernel void part1(__global int* a, __global int* b, __global int* c)
 {
     unsigned int i = get_global_id(0);
 
-    //c[i] = a[i] + b[i];
     c[i] = function_example(a[i], b[i]);
 }
 );
